API: Rename panel.py to signal.py
