+++ typhon/signal.py
